Work on a new server mode that takes advantage of pynbody's shared memory arrays
