+++ tangos/input_handlers/pynbody.py
@@ -9,6 +9,8 @@ from itertools import chain
 import numpy as np
 from more_itertools import always_iterable
 
+import tangos.parallel_tasks.pynbody_server.snapshot_queue
+
 from ..util import proxy_object
 
 pynbody = None # deferred import; occurs when a PynbodyInputHandler is constructed
@@ -117,11 +119,13 @@ class PynbodyInputHandler(finding.PatternBasedFileDiscovery, HandlerBase):
         elif mode=='server':
             timestep = self.load_timestep(ts_extension, mode)
             from ..parallel_tasks import pynbody_server as ps
-            return timestep.get_view(ps.ObjectSpecification(finder_id, finder_offset, object_typetag))
+            return timestep.get_view(
+                tangos.parallel_tasks.pynbody_server.snapshot_queue.ObjectSpecification(finder_id, finder_offset, object_typetag))
         elif mode=='server-partial':
             timestep = self.load_timestep(ts_extension, mode)
             from ..parallel_tasks import pynbody_server as ps
-            view = timestep.get_view(ps.ObjectSpecification(finder_id, finder_offset, object_typetag))
+            view = timestep.get_view(
+                tangos.parallel_tasks.pynbody_server.snapshot_queue.ObjectSpecification(finder_id, finder_offset, object_typetag))
             load_index = view['remote-index-list']
             logger.info("Partial load %r, taking %d particles", ts_extension, len(load_index))
             f = pynbody.load(self._extension_to_filename(ts_extension), take=load_index)
